Time: 58 ms (36.36%), Space: 27 MB (81.82%) - LeetHub

+++ 0576-out-of-boundary-paths/0576-out-of-boundary-paths.cs
@@ -1,13 +1,13 @@
 public class Solution {
     int M;    
     int N;
-    long[,,] Memo;
+    int[,,] Memo;
     const long MOD = 1000000007;
     
     public int FindPaths(int m, int n, int maxMove, int startRow, int startColumn) {
         M = m;
         N = n;
-        Memo = new long[m, n, maxMove];
+        Memo = new int[m, n, maxMove];
         for (int i = 0; i < m; ++i)
         {
             for (int j = 0; j < n; ++j)
@@ -20,10 +20,10 @@ public class Solution {
         }
         
         
-        return (int)GetPathsRecursively(startRow, startColumn, maxMove);
+        return GetPathsRecursively(startRow, startColumn, maxMove);
     }
     
-    public long GetPathsRecursively(int r, int c, int stepLeft)
+    public int GetPathsRecursively(int r, int c, int stepLeft)
     {
         if (r < 0 || M == r || c < 0 || N == c)
             return 1;
@@ -33,11 +33,11 @@ public class Solution {
         --stepLeft;
         if (Memo[r, c, stepLeft] != -1)
             return Memo[r, c, stepLeft];
-        long result = GetPathsRecursively(r+1, c, stepLeft) +   
+        long result = (long) GetPathsRecursively(r+1, c, stepLeft) +   
                     GetPathsRecursively(r-1, c, stepLeft) +
                     GetPathsRecursively(r, c+1, stepLeft) +
                     GetPathsRecursively(r, c-1, stepLeft);
-        Memo[r, c, stepLeft] = result % MOD;
+        Memo[r, c, stepLeft] = (int)(result % MOD);
         return Memo[r, c, stepLeft];
     }
 }
